Fix 32 bits variables from 64 bits

+++ radixSort.h
@@ -6,24 +6,24 @@
     TYPE *repetitionsArray = (TYPE*)AllocatorCalloc(NATURAL_NUMBERS,TYPE_DATA);             \
     TYPE *temporaryArray = (TYPE*)AllocatorCalloc(SIZE_ARRAY,TYPE_DATA);                    \
     TYPE maxValue = *((TYPE*)ArrayFindMaximumValue(ARRAY,SIZE_ARRAY,TYPE_DATA));                \
-    int64_t divisor = 1;                                                                    \
+    int32_t divisor = 1;                                                                    \
                                                                                             \
     while( (maxValue/divisor) > 0)                                                          \
     {                                                                                       \
         ArrayClear(repetitionsArray,NATURAL_NUMBERS,TYPE_DATA);                             \
-        for(int64_t i = 0; i < SIZE_ARRAY; i++)                                             \
+        for(int32_t i = 0; i < SIZE_ARRAY; i++)                                             \
         {                                                                                   \
-            repetitionsArray[ (int64_t)( ((TYPE*)ARRAY)[i] / divisor) % 10]++;              \
+            repetitionsArray[ (int32_t)( ((TYPE*)ARRAY)[i] / divisor) % 10]++;              \
         }                                                                                   \
                                                                                             \
-        for(int64_t i = 1; i < NATURAL_NUMBERS; i++)                                        \
+        for(int32_t i = 1; i < NATURAL_NUMBERS; i++)                                        \
         {                                                                                   \
             repetitionsArray[i] += repetitionsArray[i-1];                                   \
         }                                                                                   \
                                                                                             \
-        for(int64_t i = (SIZE_ARRAY - 1); i >= 0; i--)                                      \
+        for(int32_t i = (SIZE_ARRAY - 1); i >= 0; i--)                                      \
         {                                                                                   \
-            int64_t aux = --repetitionsArray[ (int64_t)(((TYPE*)ARRAY)[i] / divisor) % 10]; \
+            int32_t aux = --repetitionsArray[ (int32_t)(((TYPE*)ARRAY)[i] / divisor) % 10]; \
             temporaryArray[aux] = ((TYPE*)ARRAY)[i];                                        \
         }                                                                                   \
         divisor *= 10;                                                                      \
